requirements.txt
# Core dependencies
owlready2>=0.45
anthropic>=0.18.0
sqlalchemy>=2.0.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-core>=0.1.0
pydantic>=2.0.0
pyyaml>=6.0
rdflib>=6.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# HTTP/Webhooks
httpx>=0.25.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Utilities
python-dotenv>=1.0.0
structlog>=23.2.0

# Docker
docker>=7.0.0

# MCP Protocol (using JSON-RPC for MCP implementation)
jsonrpcclient>=4.0.0
jsonrpcserver>=5.0.0
